// e2e/fixtures/gateway-poly-inferred-app/src/generated/remote.ts
// AUTO-GENERATED BY PYLON INTROSPECT

import {ID, Int, Float} from '@getcronit/pylon';

type String = string & {readonly brand?: unique symbol};
type Boolean = boolean & {readonly brand?: unique symbol};
type Number = number & {readonly brand?: unique symbol};
type Any = any & {readonly brand?: unique symbol};
type Void = void & {readonly brand?: unique symbol};
type JSONObject = object & {readonly brand?: unique symbol};
type JSON = object & {readonly brand?: unique symbol};

export interface Query {
  readonly __typename?: 'Query';
  user(id: String): User | null;
}

export interface User {
  readonly __typename: 'User';
  id: String;
  email: String;
  kind: String;
  specialty?: String;
  insuranceId?: String;
}

export interface RemoteRegistry {
  delegate: {
    'Query.user': {
      args: { id: String };
      return: User | null;
    };
  };
  types: {
    'User': User;
  };
}